+++ src/__init__.py
@@ -84,4 +84,4 @@ __all__ = [
     "TEMPLATE_MODE_PLACEHOLDER",
 ]
 
-__version__ = "6.1.0"
+__version__ = "6.2.0"